Rev C: Production Version of W2 Case but no B-Vent
Rev C is the production version of the W2 case available on getrileylink.org, but without the Barometric Vent hole in the top case. The case will fit both Medtronic and Omnipod RileyLink 1.0 boards, The Barometric vent is a custom designed component that would be impossible for a user wishing to make this case on their own to make or source at a low order quantity. Therefore, the B Vent hole was removed. The case will still work as intended, but the user will need to "burp" the air pressure from the case when putting the top on. The production version of the case including all required components is available on GetRileyLink.org.