W2 Clip Final Tooled Design
This is the final design of the W2 Clip, a belt clip for the W2 Case. The part was tooled up in injection molded polycarbonate and is available for purchase on www.getrileylink.org. For those attempting to 3D print, I would recommend printing in formlabs "tough" material for it to have any chance of surviving as well as the molded polycarbonate version.